Use a custom ClassLoader to isolate unit tests.
The tests were being run in a JVM with the agent already loaded.
This meant that tests were not isolated: they all used the same
agent instance, and hence they shared state.

By loading instrumented classes and the stateful agent classes in
a separate classloader, the tests can be isolated.  This reduces
dependencies between tests, and will allow tests to be run in
parallel.