Multiple breakpts support, bug & code format fixes
Additionally, a Settings singleton has been added to allow loading
of String -> String properties (perhaps from a config file at a
later point).

Although multiple breakpoints for taking snapshots of the current
state of a program (and outputting it to XML) are supported, there
is only one output file. This results in overwriting the output
from each previous snapshot. Possible solution: Append to file
instead of overwriting it.